# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FarmingHave API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: game
    description: Game save operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /game/save:
    get:
      operationId: getGameSave
      tags: [game]
      summary: Get game save
      description: Returns the authenticated user's game save data
      responses:
        "200":
          description: Game save data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSave"
        "404":
          description: No save found
    post:
      operationId: saveGame
      tags: [game]
      summary: Save game
      description: Save or update the authenticated user's game state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameSaveInput"
      responses:
        "200":
          description: Game saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSave"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    GameSave:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        money:
          type: integer
        plots:
          type: string
        stats:
          type: string
        unlockedAchievements:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - money
        - plots
        - stats
        - unlockedAchievements
        - updatedAt

    GameSaveInput:
      type: object
      properties:
        money:
          type: integer
        plots:
          type: string
        stats:
          type: string
        unlockedAchievements:
          type: string
      required:
        - money
        - plots
        - stats
        - unlockedAchievements