# lib/api-zod/src/generated/types/gameSave.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FarmingHave API
 * OpenAPI spec version: 0.1.0
 */

export interface GameSave {
  id: number;
  userId: string;
  money: number;
  plots: string;
  stats: string;
  unlockedAchievements: string;
  updatedAt: string;
}