# lib/api-zod/src/generated/types/gameSaveInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FarmingHave API
 * OpenAPI spec version: 0.1.0
 */

export interface GameSaveInput {
  money: number;
  plots: string;
  stats: string;
  unlockedAchievements: string;
}